chore: sync august

+++ src/components/c2d/index.ts
@@ -30,7 +30,7 @@ export async function decryptFilesObject(
     )
 
     // 3. Convert the decrypted bytes back to a string
-    const decryptedFilesString = Buffer.from(decryptedUrlBytes).toString()
+    const decryptedFilesString = Buffer.from(decryptedUrlBytes as Uint8Array).toString()
     const decryptedFileArray = JSON.parse(decryptedFilesString)
 
     console.log('decryptedFileArray: ', decryptedFileArray)
